backend/requirements.txt
fastapi
uvicorn
pymongo
pandas
scikit-learn
python-dotenv
requests
pydantic
supabase
sqlalchemy
psycopg2-binary
redis
httpx
sentence-transformers
torch
pgvector
qdrant-client
